Andrew Parise - added authorization checks to todo endpoints

+++ sample-web-app/src/main/java/org/tadalabs/sample/web/controller/TodoController.java
@@ -4,10 +4,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.tadalabs.sample.core.boundary.enter.SessionService;
 import org.tadalabs.sample.core.domain.Todo;
 import org.tadalabs.sample.adapter.web.api.TodoList;
 import org.tadalabs.sample.core.boundary.enter.TodoService;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 import javax.validation.constraints.NotBlank;
 import java.util.Optional;
@@ -18,9 +20,12 @@ public class TodoController {
 
     private final TodoService todoService;
 
+    private final SessionService sessionService;
+
     @Autowired
-    public TodoController(TodoService todoService) {
+    public TodoController(TodoService todoService, SessionService sessionService) {
         this.todoService = todoService;
+        this.sessionService = sessionService;
     }
 
     @PostMapping(produces = "application/json", consumes = "application/json")
@@ -38,9 +43,15 @@ public class TodoController {
 
     @GetMapping(value = "/", produces = "application/json")
     @ResponseBody
-    public ResponseEntity<TodoList> getAllTodos() {
+    public ResponseEntity<TodoList> getAllTodos(
+            @RequestHeader(value = "Authorization", defaultValue = "") String authorizationToken,
+            HttpServletRequest httpServletRequest) {
 
-        Optional<TodoList> optional = this.todoService.findAll();
+        if(!this.sessionService.isSessionValid(authorizationToken, httpServletRequest.getRemoteAddr())) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        }
+
+        Optional<TodoList> optional = this.todoService.findAllWithSessionId(authorizationToken);
 
         return optional.map(todoList -> new ResponseEntity<>(todoList, HttpStatus.OK))
                 .orElseGet(() -> new ResponseEntity<>(HttpStatus.NO_CONTENT));
@@ -48,9 +59,16 @@ public class TodoController {
 
     @GetMapping(value = "/{todoId}", produces = "application/json")
     @ResponseBody
-    public ResponseEntity<Todo> getTodoById(@Valid @NotBlank @PathVariable("todoId") String todoId) {
+    public ResponseEntity<Todo> getTodoById(
+            @RequestHeader(value = "Authorization", defaultValue = "") String authorizationToken,
+            HttpServletRequest httpServletRequest,
+            @Valid @NotBlank @PathVariable("todoId") String todoId) {
 
-        Optional<Todo> optionalTodo = this.todoService.getTodoById(todoId);
+        if(!this.sessionService.isSessionValid(authorizationToken, httpServletRequest.getRemoteAddr())) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        }
+
+        Optional<Todo> optionalTodo = this.todoService.getByTodoIdAndSessionId(todoId, authorizationToken);
         if(!optionalTodo.isPresent()) {
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         }
@@ -59,9 +77,16 @@ public class TodoController {
                 .body(optionalTodo.get());
     }
 
-    @PutMapping(value = "/{todoId}", produces = "application/json")
+    @PutMapping(value = "/", produces = "application/json")
     @ResponseBody
-    public ResponseEntity updateTodo(@PathVariable String todoId, @Valid @RequestBody Todo request) {
+    public ResponseEntity updateTodo(
+            @RequestHeader(value = "Authorization", defaultValue = "") String authorizationToken,
+            HttpServletRequest httpServletRequest,
+            @Valid @RequestBody Todo request) {
+
+        if(!this.sessionService.isSessionValid(authorizationToken, httpServletRequest.getRemoteAddr())) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        }
 
         Optional<Todo> optionalTodo = this.todoService.updateTodo(request);
         if(!optionalTodo.isPresent()) {
@@ -73,7 +98,14 @@ public class TodoController {
 
     @DeleteMapping(value = "/{todoId}", produces = "application/json")
     @ResponseBody
-    public ResponseEntity deleteTodo(@PathVariable String todoId) {
+    public ResponseEntity deleteTodo(
+            @RequestHeader(value = "Authorization", defaultValue = "") String authorizationToken,
+            HttpServletRequest httpServletRequest,
+            @PathVariable String todoId) {
+
+        if(!this.sessionService.isSessionValid(authorizationToken, httpServletRequest.getRemoteAddr())) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        }
 
         this.todoService.removeTodo(todoId);
         return ResponseEntity.status(HttpStatus.OK).build();
